tech-ai-bot/requirements.txt
google-generativeai>=0.7.0
requests>=2.31.0
tweepy>=4.14.0
python-dotenv>=1.0.0
tenacity>=8.2.0